refactor: 프롬프트 수정, 패키지 이동으로 ENUM 재명시
