+++ src/main/java/nbc/devmountain/domain/ai/service/AiService.java
@@ -1,4 +1,3 @@
-// src/main/java/nbc/devmountain/domain/ai/service/AiService.java
 package nbc.devmountain.domain.ai.service;
 
 import java.util.Collections;
@@ -9,7 +8,6 @@ import java.util.regex.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-import org.springframework.ai.chat.messages.MessageType;
 import org.springframework.ai.chat.messages.SystemMessage;
 import org.springframework.ai.chat.messages.UserMessage;
 import org.springframework.ai.chat.model.ChatModel;
@@ -22,7 +20,8 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import nbc.devmountain.domain.ai.dto.RecommendationDto;
-import nbc.devmountain.domain.chat.chatmessage.dto.response.ChatMessageResponse;
+import nbc.devmountain.domain.chat.dto.ChatMessageResponse;
+import nbc.devmountain.domain.chat.model.MessageType;
 
 @Service
 @AllArgsConstructor
@@ -37,21 +36,10 @@ public class AiService {
 				"""
 				너는 주어진 정보를 바탕으로 강의를 추천하는 교육 큐레이터 AI야.
 				- 사용자의 질문과 제공된 '유사한 강의 정보'를 바탕으로 가장 적절한 강의를 최대 3개 추천해줘.
-				- 각 강의는 title, url, level, thumbnailUrl 을 포함해야 해.
+				- 각 강의는 title, description, instructor, level, thumbnailUrl 을 포함해야 해.
 				- 응답은 반드시 JSON 형식으로만 해야하며, 절대로 JSON 객체 외의 다른 텍스트(예: 설명, 인사)를 포함하면 안돼.
 				- 만약 추천할 강의가 없다면, recommendations 배열을 비워서 보내줘. 예: {"recommendations": []}
-				- 응답 예시:
-				{
-				"recommendations": [
-				  {
-				    "title": "자바 입문",
-				    "description": "...",
-				    "instructor": "...",
-				    "level": "초급",
-				    "thumbnailUrl": "..."
-				  }
-				 ]
-				}
+				- 응답 예시: {"recommendations": [{"title": "스프링 입문", "description": "...", "instructor": "...", "level": "초급", "thumbnailUrl": "some_url.jpg" }]}
 				"""
 				:
 				"""
@@ -78,7 +66,7 @@ public class AiService {
 			return ChatMessageResponse.builder()
 				.message(rawAiResponse)
 				.isAiResponse(true)
-				.messageType(ChatMessageResponse.MessageType.CHAT)
+				.messageType(MessageType.CHAT)
 				.build();
 		}
 
@@ -90,10 +78,15 @@ public class AiService {
 		}
 
 		try {
-			JsonNode root =objectMapper.readTree(pureJson);
+			JsonNode root = objectMapper.readTree(pureJson);
 			JsonNode recNode = root.get("recommendations");
-			List<RecommendationDto> recommendations = objectMapper.readValue(
-				recNode.toString(), objectMapper.getTypeFactory().constructCollectionType(List.class, RecommendationDto.class));
+
+			if (recNode == null || !recNode.isArray()) {
+				log.warn("[AiService] AI 응답에 'recommendations' 배열이 없거나 형식이 올바르지 않음.");
+				return createErrorResponse("AI가 올바른 형식의 응답을 생성하지 못했습니다. 다시 시도해주세요.");
+			}
+			List<RecommendationDto> recommendations = objectMapper.convertValue(recNode, // 수정된 부분
+				objectMapper.getTypeFactory().constructCollectionType(List.class, RecommendationDto.class));
 
 			if (recommendations == null || recommendations.isEmpty()) {
 				log.info("[AiService] AI가 추천할 강의를 찾지 못함.");
@@ -105,12 +98,12 @@ public class AiService {
 				.message(null)
 				.recommendations(recommendations)
 				.isAiResponse(true)
-				.messageType(ChatMessageResponse.MessageType.RECOMMENDATION)
+				.messageType(MessageType.RECOMMENDATION)
 				.build();
 
 		} catch (JsonProcessingException e) {
-			log.error("[AiService] AI 응답 파싱 실패!\n원본: {}\n추출된 JSON: {}\n에러: {}",
-				rawAiResponse, pureJson, e.toString());
+			log.error("[AiService] AI 응답 파싱 실패!\n원본: {}\n추출된 JSON: {}\n에러: {}\n{}",
+				rawAiResponse, pureJson, e.toString(), e.getMessage());
 			return createErrorResponse("AI 응답을 처리하는 중 오류가 발생했습니다. 잠시 후 다시 시도해주세요.");
 		}
 	}
@@ -133,7 +126,7 @@ public class AiService {
 			.message(errorMessage)
 			.recommendations(Collections.emptyList())
 			.isAiResponse(true)
-			.messageType(ChatMessageResponse.MessageType.ERROR)
+			.messageType(MessageType.ERROR)
 			.build();
 	}
 }
